Summary: Merge Gerard's AIX signed and datatype size patches. Keywords:
Patches applied:

 * gerard@eviston.net--squid/squid--aix-fixes--3.0--patch-5
   typecast ntohl() in wccp.cc

 * ports@eviston.net--aix/squid--aix-fixes--3.0--patch-3
   debug() ==> debugs() round 1

 * ports@eviston.net--aix/squid--aix-fixes--3.0--patch-4
   casts and format strings

 * ports@eviston.net--aix/squid--aix-fixes--3.0--patch-5
   debug() => debugs() round 2: peer_digest.cc

 * ports@eviston.net--aix/squid--aix-fixes--3.0--patch-6
   casts in stat.cc

 * ports@eviston.net--aix/squid--aix-fixes--3.0--patch-7
   debug() => debugs() round 3

 * ports@eviston.net--aix/squid--aix-fixes--3.0--patch-8
   cast for ntohl() in wccp.cc

 * ports@eviston.net--aix/squid--aix-fixes--3.0--patch-9
   casts in tools.cc

+++ src/tools.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: tools.cc,v 1.239 2003/08/03 10:05:12 robertc Exp $
+ * $Id: tools.cc,v 1.240 2003/09/29 10:24:02 robertc Exp $
  *
  * DEBUG: section 21    Misc Functions
  * AUTHOR: Harvest Derived
@@ -157,23 +157,23 @@ dumpMallocStats(void)
 
     fprintf(debug_log, "Memory usage for %s via mallinfo():\n", appname);
 
-    fprintf(debug_log, "\ttotal space in arena:  %6d KB\n",
-            mp.arena >> 10);
+    fprintf(debug_log, "\ttotal space in arena:  %6ld KB\n",
+            (long)mp.arena >> 10);
 
-    fprintf(debug_log, "\tOrdinary blocks:       %6d KB %6d blks\n",
-            mp.uordblks >> 10, mp.ordblks);
+    fprintf(debug_log, "\tOrdinary blocks:       %6ld KB %6ld blks\n",
+            (long)mp.uordblks >> 10, (long)mp.ordblks);
 
-    fprintf(debug_log, "\tSmall blocks:          %6d KB %6d blks\n",
-            mp.usmblks >> 10, mp.smblks);
+    fprintf(debug_log, "\tSmall blocks:          %6ld KB %6ld blks\n",
+            (long)mp.usmblks >> 10, (long)mp.smblks);
 
-    fprintf(debug_log, "\tHolding blocks:        %6d KB %6d blks\n",
-            mp.hblkhd >> 10, mp.hblks);
+    fprintf(debug_log, "\tHolding blocks:        %6ld KB %6ld blks\n",
+            (long)mp.hblkhd >> 10, (long)mp.hblks);
 
-    fprintf(debug_log, "\tFree Small blocks:     %6d KB\n",
-            mp.fsmblks >> 10);
+    fprintf(debug_log, "\tFree Small blocks:     %6ld KB\n",
+            (long)mp.fsmblks >> 10);
 
-    fprintf(debug_log, "\tFree Ordinary blocks:  %6d KB\n",
-            mp.fordblks >> 10);
+    fprintf(debug_log, "\tFree Ordinary blocks:  %6ld KB\n",
+            (long)mp.fordblks >> 10);
 
     t = mp.uordblks + mp.usmblks + mp.hblkhd;
